requirements.txt
Flask==3.1.0
Requests==2.32.3
mutagen==1.47.0
python-slugify==8.0.4
ffmpeg-python==0.2.0
waitress==3.0.2